Split RepoModule to NetworkModule and RepoModule
