+++ app/src/main/java/com/edu/kotlindbnetwork/AppComponent.kt
@@ -2,6 +2,7 @@ package com.edu.kotlindbnetwork
 
 import com.edu.kotlindbnetwork.data.db.Database
 import com.edu.kotlindbnetwork.data.network.APIService
+import com.edu.kotlindbnetwork.modules.NetworkModule
 import com.edu.kotlindbnetwork.modules.RepoModule
 import com.edu.kotlindbnetwork.repo.UserRepo
 import com.edu.kotlindbnetwork.ui.MainActivity
@@ -10,8 +11,9 @@ import com.edu.kotlindbnetwork.ui.UserProfileFragment
 
 import dagger.Component
 import retrofit2.Retrofit
+import javax.inject.Named
 
-@Component(modules = [RepoModule::class])
+@Component(modules = [RepoModule::class, NetworkModule::class])
 interface AppComponent {
 
     fun injectsMainActivity(mainActivity: MainActivity)
@@ -20,7 +22,7 @@ interface AppComponent {
 
     fun getUserRepo(): UserRepo
     fun getDatabase(): Database
-    fun getApiService(): APIService
+    @Named(Consts.MODULE_TAG_API_SERVICE) fun getApiService(): APIService
     fun getRetrofit(): Retrofit
 
 }
